requirements.txt
pyyaml
numpy
easydict==1.9.0
opencv-python
matplotlib
pandas
pytorch-lightning==1.2.9

# tabulate
# tqdm
# scikit-image==0.17.2
# scikit-learn==0.24.2
# tensorflow
# joblib
# albumentations==0.5.2
# hydra-core==1.1.0
# kornia==0.5.0
# webdataset
# packaging
# scikit-learn==0.24.2
# wldhx.yadisk-direct
# mahotas
# simpleITK
# theano